vdk-control-cli: Clean up auth leftovers
As we adopt the new stand-alone authentication plugin (`vdk-control-api-auth`),
we can remove the authentication-related code from `vdk-control-cli`, as it is
no longer needed.

This change cleans up all authentication-related code from `vdk-control-cli`, and
fully adopts `vdk-control-api-auth` as the de-facto authentication library.

Testing Done: Unit tests
